Changes on how Logging is done

+++ projects/project_1/src/utils.c
@@ -1,10 +1,10 @@
 #include "utils.h"
 
 /* Calculates time (in ms) since the start of the program */
-int delta() {
+double delta() {
   struct timespec end;
   clock_gettime(CLOCK_MONOTONIC_RAW, &end);
-  return (end.tv_sec - start.tv_sec) * 1000 + (end.tv_nsec - start.tv_nsec) / 1000000;
+  return (end.tv_sec - start.tv_sec) * 1000.0 + (end.tv_nsec - start.tv_nsec) / 1000000.0;
 }
